Create JIMM v4 class.
