+++ src/juju/jimm/index.ts
@@ -1,10 +1,12 @@
 import JIMMV3 from "./JIMMV3";
+import JIMMV4 from "./JIMMV4";
 
 export * as JIMMV3 from "./JIMMV3";
+export * as JIMMV4 from "./JIMMV4";
 
 const JIMM = {
   name: "JIMM",
-  versions: [JIMMV3],
+  versions: [JIMMV3, JIMMV4],
 };
 
 export default JIMM;
