+++ src/juju/types.ts
@@ -13,7 +13,7 @@ import type ModelManagerV9 from "@canonical/jujulib/dist/api/facades/model-manag
 import type { ModelSLAInfo } from "@canonical/jujulib/dist/api/facades/model-manager/ModelManagerV9";
 import type PingerV1 from "@canonical/jujulib/dist/api/facades/pinger/PingerV1";
 
-import type JIMMV3 from "./jimm/JIMMV3";
+import type JIMMV4 from "./jimm/JIMMV4";
 // See https://github.com/juju/juju/blob/main/rpc/params/multiwatcher.go
 // for the Juju types for the AllWatcher responses.
 
@@ -347,7 +347,7 @@ export type Facades = {
   controller?: ControllerV9;
   modelManager?: ModelManagerV9;
   pinger?: PingerV1;
-  jimM?: InstanceType<typeof JIMMV3>;
+  jimM?: InstanceType<typeof JIMMV4>;
 };
 
 export type ConnectionWithFacades = Connection & {
